Move interface to types.ts
