+++ packages/caravan-clients/src/client.ts
@@ -18,6 +18,7 @@ import {
   bitcoindWalletInfo,
 } from "./wallet";
 import BigNumber from "bignumber.js";
+import { FeeRatePercentile, Transaction, UTXO } from "./types";
 
 export class BlockchainClientError extends Error {
   constructor(message) {
@@ -26,56 +27,10 @@ export class BlockchainClientError extends Error {
   }
 }
 
-export interface UTXO {
-  txid: string;
-  vout: number;
-  value: number;
-  status: {
-    confirmed: boolean;
-    block_time: number;
-  };
-}
-
-export interface Transaction {
-  txid: string;
-  vin: Input[];
-  vout: Output[];
-  size: number;
-  weight: number;
-  fee: number;
-  isSend: boolean;
-  amount: number;
-  blocktime: number;
-}
-
-interface Input {
-  txid: string;
-  vout: number;
-  sequence: number;
-}
-
-interface Output {
-  scriptPubkeyHex: string;
-  scriptPubkeyAddress: string;
-  value: number;
-}
-
 export enum ClientType {
   PRIVATE = "private",
   BLOCKSTREAM = "blockstream",
   MEMPOOL = "mempool",
-}
-
-export interface FeeRatePercentile {
-  avgHeight: number;
-  timestamp: number;
-  avgFee_0: number;
-  avgFee_10: number;
-  avgFee_25: number;
-  avgFee_50: number;
-  avgFee_75: number;
-  avgFee_90: number;
-  avgFee_100: number;
 }
 
 const delay = () => {
