+++ packages/caravan-clients/src/index.ts
@@ -1,3 +1,3 @@
 export { bitcoindImportDescriptors } from "./wallet";
 export { BlockchainClient, ClientType } from "./client";
-export type { UTXO, Transaction, FeeRatePercentile } from "./client";
+export type { UTXO, Transaction, FeeRatePercentile } from "./types";
